Rename tags.ts -> token-node-defintions.ts
